Started using PHP to import navbar
I was using Server-Side Includes, but this was messing up my PHP for
determining my current page (so that it could be highlighted in the
navbar). I am now using PHP for this purpose as it does the job just as
well, but additionally enables the rest of my PHP code to work in a
simple manner.